refactor : added useDocumentTitle hook to all the pages
