+++ src/pages/Home/home.jsx
@@ -1,9 +1,12 @@
 import React from 'react';
 import { Header, HeroImage, Footer, WhyUsSection, Categories, Section } from '../../components'
-
+import useDocumentTitle from '../../utils/hooks/useDocumentTitle'
 import './home.css'
 
 function HomePage(){
+
+    useDocumentTitle('Paws & Beyond | Home')
+
     return(
         <div className="home-bg">
             <Header />
