+++ src/pages/PageNotFound/404page.jsx
@@ -1,6 +1,10 @@
 import React from 'react';
+import useDocumentTitle from '../../utils/hooks/useDocumentTitle'
 
 function PageNotFound(){
+
+    useDocumentTitle('Paws & Beyond | 404')
+
     return(
         <div>
             404 page not found :(
