+++ src/pages/WishList/wishList.jsx
@@ -4,8 +4,11 @@ import './wishList.css'
 import { Header, Footer, ProductCard, EmptySection } from '../../components'
 import { emptyWishList } from '../../assets/images';
 import { useWishList } from '../../context'
+import useDocumentTitle from '../../utils/hooks/useDocumentTitle';
 
 function WishListPage(){
+
+    useDocumentTitle('Paws & Beyond | Wishlist')
 
     const { wishList } = useWishList()
     let productCount = wishList.length
